+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Print how long the certificate is valid for when downloaded.
 - Fall back to a wider range of default identities.
 - Add ability to disable opening the browser for authentication. Pass `--browser=false` to the `clifton auth` or set `open_browser = false` in the config file.
+- Warn if using an unencrypted private key.
 
 ### Fixed
 - Allow tilde in arguments.